Remove custom allocator, fallback to using global allocator
This is done because it has become incredibly cumbersome to work with
so many lifetimes in an effective way, and has gotten too much in the
way of rapid prototyping. The measured performance gains from using a
bump allocator are not worth the mental overhead and visual noise
produced by 'ast and 'alloc lifetimes leaking all over the code, as well
as questionable uses of unsafe code to get around the borrow checker.