Tokeniser: Support emitting errors from 'char' and 'string' functions.
- Now support processing unicode literals in the formn of `\u{..}`

- `TokenKind` generating functions may return a `TokenResult` in the event that they
encounter an unrecoverable tokenisation state. This mainly occurs when processing
string and character literals that have pre-defined rules about how they are expressed.
Later, `TokenError` will be castable to a  `ParseError` so that we can use our
reporting system to report the issue.

- fix missing panic messages. Panic handler now supports formatted messages to
the macro since it didn't by default. Given that this behaviour requires rust
nightly for some ungodly reason (which is pretty strange), the project also depends
on `#![feature(panic_info_message)]` now as well.